Trace Interpreter: fix glWaitSync handling
* GL_TIMEOUT_IGNORED was being parsed as an integer (resulting in 0)
* GL_TIMEOUT_IGNORED handling in GLenumToString was incorrect due to
  too small data type (needs 64 bits)

Enables asphalt_9 in retraced interpreter tests.

Bug: b/276742336

+++ src/common/gl_enum_utils.cpp
@@ -123,9 +123,9 @@ const char *GLinternalFormatToString(unsigned int format)
     return GLenumToString(gl::GLESEnum::InternalFormat, format);
 }
 
-unsigned int StringToGLbitfield(const char *str)
+uint64_t StringToGLbitfield(const char *str)
 {
-    unsigned int value = 0;
+    uint64_t value = 0;
     std::vector<std::string> strings =
         angle::SplitString(str, " |", angle::WhitespaceHandling::TRIM_WHITESPACE,
                            angle::SplitResult::SPLIT_WANT_NONEMPTY);
